feat/refactor: Декомпозиция offer-page/Изменена файловая структура
