+++ src/components/locations/locations-item/locations-item.tsx
