+++ src/components/offer/offer-info/offer-info-components/info-host/info-host-components/host-description/host-description.tsx
@@ -1,0 +1,19 @@
+
+function HostDescription():JSX.Element{
+  return (
+    <div className="offer__description">
+      <p className="offer__text">
+                A quiet cozy and picturesque that hides behind a a river by the
+                unique lightness of Amsterdam. The building is green and from
+                18th century.
+      </p>
+      <p className="offer__text">
+                An independent House, strategically located between Rembrand
+                Square and National Opera, but where the bustle of the city
+                comes to rest in this alley flowery and colorful.
+      </p>
+    </div>
+  );
+}
+
+export default HostDescription;
