+++ src/components/offer/offer-info/offer-info-components/info-host/info-host-components/host-user/host-user.tsx
@@ -1,0 +1,20 @@
+
+function HostUser():JSX.Element{
+  return (
+    <div className="offer__host-user user">
+      <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
+        <img
+          className="offer__avatar user__avatar"
+          src="img/avatar-angelina.jpg"
+          width={74}
+          height={74}
+          alt="Host avatar"
+        />
+      </div>
+      <span className="offer__user-name">Angelina</span>
+      <span className="offer__user-status">Pro</span>
+    </div>
+  );
+}
+
+export default HostUser;
