+++ src/components/offer/offer-info/offer-info-components/info-name/info-name.tsx
@@ -1,0 +1,14 @@
+import BookmarkButton from '../../../../bookmark-button/bookmark-button';
+
+function InfoName():JSX.Element{
+  return (
+    <div className="offer__name-wrapper">
+      <h1 className="offer__name">
+              Beautiful &amp; luxurious studio at great location
+      </h1>
+      <BookmarkButton bookmarkClass='offer'/>
+    </div>
+  );
+}
+
+export default InfoName;
