+++ src/components/offer/offer-info/offer-info-components/info-price/info-price.tsx
@@ -1,0 +1,11 @@
+
+function InfoPrice():JSX.Element{
+  return (
+    <div className="offer__price">
+      <b className="offer__price-value">€120</b>
+      <span className="offer__price-text">&nbsp;night</span>
+    </div>
+  );
+}
+
+export default InfoPrice;
