+++ src/components/offer/offer-info/offer-info-components/info-reviews/info-reviews-components/reviews-item/reviews-item-components/reviews-user/reviews-user.tsx
@@ -1,0 +1,19 @@
+
+function ReviewsUser():JSX.Element {
+  return (
+    <div className="reviews__user user">
+      <div className="reviews__avatar-wrapper user__avatar-wrapper">
+        <img
+          className="reviews__avatar user__avatar"
+          src="img/avatar-max.jpg"
+          width={54}
+          height={54}
+          alt="Reviews avatar"
+        />
+      </div>
+      <span className="reviews__user-name">Max</span>
+    </div>
+  );
+}
+
+export default ReviewsUser;
