+++ src/components/offer/offer-info/offer-info-components/info-reviews/info-reviews-components/reviews-item/reviews-item.tsx
@@ -1,0 +1,13 @@
+import ReviewsInfo from './reviews-item-components/reviews-info/reviews-info';
+import ReviewsUser from './reviews-item-components/reviews-user/reviews-user';
+
+function ReviewsItem():JSX.Element{
+  return (
+    <li className="reviews__item">
+      <ReviewsUser/>
+      <ReviewsInfo/>
+    </li>
+  );
+}
+
+export default ReviewsItem;
