+++ src/components/offer/offer-info/offer-info-components/info-reviews/info-reviews-components/reviews-list/reviews-list.tsx
@@ -1,0 +1,11 @@
+import ReviewsItem from '../reviews-item/reviews-item';
+
+function ReviewsList():JSX.Element{
+  return (
+    <ul className="reviews__list">
+      <ReviewsItem/>
+    </ul>
+  );
+}
+
+export default ReviewsList;
