+++ src/components/place-cards/place-cards-list/place-cards-list.tsx
