+++ src/components/sort-offers/sort-offers-list/sort-offers-list.tsx
@@ -1,6 +1,6 @@
 import { nanoid } from 'nanoid';
 import SortOffersItem from '../sort-offers-item/sort-offers-item';
-import { SortOptions } from '../../const';
+import { SortOptions } from '../../../const';
 
 function SortOffersList():JSX.Element {
   const sortOptionsList = Object.values(SortOptions).map((option) => <SortOffersItem sortOption={option} isActive={option === 'Popular'} key={nanoid()}/>);
