+++ src/index.tsx
@@ -13,6 +13,7 @@ root.render(
   <React.StrictMode>
     <App
       cardsCount = {Settings.CardsCount}
+      galleryImagesCount={Settings.GalleryImagesCount}
       offersCount = {MockData.length}
     />
   </React.StrictMode>
