+++ src/pages/login-page/login-page.tsx
@@ -1,4 +1,4 @@
-import Header from '../../components/header/header';
+import Header from '../../components/header/header-general/header-general';
 
 
 function LoginPage(): JSX.Element {
