+++ src/pages/offer-page/offer-page.tsx
@@ -1,18 +1,20 @@
-import Header from '../../components/header/header';
-import Map from '../../components/map/map';
+import Header from '../../components/header/header-general/header-general';
+import Map from '../../components/common/map/map';
 import NearPlaces from '../../components/near-places/near-places';
-import OfferGallery from '../../components/offer-gallery/offer-gallery';
-import OfferInfo from '../../components/offer-info/offer-info';
+import OfferGallery from '../../components/offer/offer-gallery/offer-gallery';
+import OfferInfo from '../../components/offer/offer-info/offer-info';
 
+type OfferPageProps = {
+  galleryImagesCount: number;
+}
 
-function OfferPage(): JSX.Element{
+function OfferPage({galleryImagesCount}: OfferPageProps): JSX.Element{
   return (
     <div className="page">
       <Header/>
-
       <main className="page__main page__main--offer">
         <section className="offer">
-          <OfferGallery/>
+          <OfferGallery GalleryImagesCount={galleryImagesCount}/>
           <OfferInfo/>
           <Map mapClass='offer'/>
         </section>
@@ -21,8 +23,6 @@ function OfferPage(): JSX.Element{
         </div>
       </main>
     </div>
-
-
   );
 }
